Adding syntax highlighter
# Description :

Add the possibility to use synthax highlighting for code block that use the "fenced" syntax :
```
    ```
    I'm a block of code
    ```
    ```QML
    I'm a (QML) highlighted block of code
    ```
```

User can use the following highlighters:
- [KSyntaxHighlighter](https://invent.kde.org/frameworks/syntax-highlighting)
- [chroma](https://github.com/alecthomas/chroma)
- [pygments](https://pygments.org/docs/styles/)

But other could easily be added.

This also expose the backed in styles of the highlighter to the user.

![image](/uploads/ef93685ab93d758aee3df90d696f6ab1/image.png)

This would be considered a "plugin" and is disabled by default.