Adding note linking
This solve #4 by adding the possibility to link notes between each other, and even do so at a header level.

Links can be created using the following synthax: `[[ Category/Group/Note : some Header | Displayed name ]]` <br>
An entry for as been added to the cheatsheet.

The path can have multiple form:
- Shown in the exemple: speaks for itself (works with or without a leading '/')
- Single word: the name of the note in the same Category/Group as the current one
- ./SomeName: same as above 
- /Category/Note: a note directly inside a category

If no "Displayed name" is provided, the link will have the note's name inside the display.

A header (of any level) can be specified or not. If no level is specified ("## some header") it will be assumed to be of level 1.
<br>
<br>

The noteMapper will save its state between each session inside a document called `notesMap.json` which is stored inside storage. 
This document allow the noteMapper to know which headers exists in which notes even if those haven't been opened in the current session. 
If the mapper doesn't have any info about the note, it will check by itself, assuming the path is correct.
<br>
<br>

Linked note can be acces in 2 ways:

<details><summary>from the preview</summary>
![note_linking](/uploads/34cf0299c4d8a3c45509469c36bcffbe/note_linking.gif)
</details>

<details><summary>from the "note map"</summary>
![note_mapper](/uploads/8971a000b8c34d82007193d6997d1d2c/note_mapper.gif)
</details>
<br>

Since this feature is pretty outside of the normal Markdown scope, it is considered a 'plugin' and can be enabled/disabled at any time in the settings. It is disabled by default.