Change the way TextToolBar actions are setup
All actions now comes from a separated file
'sharedComponents/ActionsList.qml'.
This file contains a list of KAction.
KAction is a Kirigami.Action with 2 property:
- actionName, to easily identify the action
- triggerFunction, a function that will be triggered 'onTriggered'

TextToolBar now only contains the map of actionName(key) and trigger function(value)
that will be apply to each Action based on actionName.

This enable us to use the same action across multiple file with a different
triggered function based on the usecase.